Add running as a daemon, remove running as script
The Rook server can now be run as a daemon and the functionality to
kill the server process by token has been removed as it was not
reliable.

Starting the server from a script is also removed as R will crash when
a handle (just a string) from another session.